Refactor tests to use unsigned char and unsigned long types
- Replaced instances of uint8_t with unsigned char for better compatibility.
- Changed size_t to unsigned long in various test cases to ensure consistency.
- Updated memory allocation and buffer handling to accommodate type changes.
- Enhanced assertions to handle edge cases, such as empty inputs and memory comparisons.
- Removed unnecessary includes and added comments for clarity.
- Ensured all test cases maintain functionality while improving type safety and readability.

+++ include/utils/file_list.h
@@ -1,16 +1,12 @@
 /**
  * file_list.h - Header file for file list management using linked lists
- * 
- * This header file provides the data structures and function prototypes
+ * * This header file provides the data structures and function prototypes
  * for managing a list of files using a linked list implementation.
- * 
- * Allowed libraries: stdio.h, stdlib.h, string.h, math.h
+ * * Allowed libraries: stdio.h, stdlib.h, string.h, math.h
  */
 
 #ifndef FILE_LIST_H
 #define FILE_LIST_H
-
-#include <stddef.h>  /* For size_t */
 
 /**
  * Maximum length of a filename in the file list
@@ -21,49 +17,48 @@
  * File entry structure
  * Contains information about a processed file
  */
-typedef struct file_entry {
-    char filename[FILE_LIST_MAX_FILENAME];  /* Filename */
-    unsigned long sequence_num;             /* Processing sequence number */
-    size_t original_size;                   /* Original file size */
-    size_t processed_size;                  /* Size after processing */
-    struct file_entry *next;                /* Pointer to next entry */
+typedef struct file_entry
+{
+    char filename[FILE_LIST_MAX_FILENAME]; /* Filename */
+    unsigned long sequence_num;            /* Processing sequence number */
+    unsigned long original_size;           /* Original file size - Replaced size_t with unsigned long */
+    unsigned long processed_size;          /* Size after processing - Replaced size_t with unsigned long */
+    struct file_entry *next;               /* Pointer to next entry */
 } file_entry_t;
 
 /**
  * File list structure
  * Contains the head and tail of the linked list
  */
-typedef struct {
-    file_entry_t *head;  /* Pointer to the first entry */
-    file_entry_t *tail;  /* Pointer to the last entry */
-    size_t count;        /* Number of entries in the list */
-    unsigned long next_sequence_num;  /* Next sequence number to assign */
+typedef struct
+{
+    file_entry_t *head;              /* Pointer to the first entry */
+    file_entry_t *tail;              /* Pointer to the last entry */
+    unsigned long count;             /* Number of entries in the list - Replaced size_t with unsigned long */
+    unsigned long next_sequence_num; /* Next sequence number to assign */
 } file_list_t;
 
 /**
  * Initialize a file list
- * 
- * @param list Pointer to the file list to initialize
+ * * @param list Pointer to the file list to initialize
  * @return     0 on success, -1 on failure
  */
 int file_list_init(file_list_t *list);
 
 /**
  * Add a file to the list
- * 
- * @param list           Pointer to the file list
+ * * @param list           Pointer to the file list
  * @param filename       Filename to add
- * @param original_size  Original file size
- * @param processed_size Size after processing
+ * @param original_size  Original file size - Replaced size_t with unsigned long
+ * @param processed_size Size after processing - Replaced size_t with unsigned long
  * @return               0 on success, -1 on failure
  */
-int file_list_add(file_list_t *list, const char *filename, 
-                 size_t original_size, size_t processed_size);
+int file_list_add(file_list_t *list, const char *filename,
+                  unsigned long original_size, unsigned long processed_size);
 
 /**
  * Find a file in the list
- * 
- * @param list     Pointer to the file list
+ * * @param list     Pointer to the file list
  * @param filename Filename to find (can be partial)
  * @return         Pointer to file entry if found, NULL if not found
  */
@@ -71,18 +66,16 @@ file_entry_t *file_list_find(file_list_t *list, const char *filename);
 
 /**
  * Get the most recent files in the list
- * 
- * @param list   Pointer to the file list
- * @param count  Maximum number of entries to retrieve
+ * * @param list   Pointer to the file list
+ * @param count  Maximum number of entries to retrieve - Replaced size_t with unsigned long
  * @param result Array to store the results (must be pre-allocated)
- * @return       Number of entries retrieved
+ * @return       Number of entries retrieved - Replaced size_t with unsigned long
  */
-size_t file_list_get_recent(file_list_t *list, size_t count, file_entry_t **result);
+unsigned long file_list_get_recent(file_list_t *list, unsigned long count, file_entry_t **result);
 
 /**
  * Save the file list to a file
- * 
- * @param list     Pointer to the file list
+ * * @param list     Pointer to the file list
  * @param filename Filename to save to
  * @return         0 on success, -1 on failure
  */
@@ -90,8 +83,7 @@ int file_list_save(file_list_t *list, const char *filename);
 
 /**
  * Load the file list from a file
- * 
- * @param list     Pointer to the file list
+ * * @param list     Pointer to the file list
  * @param filename Filename to load from
  * @return         0 on success, -1 on failure
  */
@@ -99,15 +91,13 @@ int file_list_load(file_list_t *list, const char *filename);
 
 /**
  * Free all memory used by the file list
- * 
- * @param list Pointer to the file list
+ * * @param list Pointer to the file list
  */
 void file_list_free(file_list_t *list);
 
 /**
  * Print the file list to stdout
- * 
- * @param list Pointer to the file list
+ * * @param list Pointer to the file list
  */
 void file_list_print(file_list_t *list);
